DarwinG-cdp/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
eth-account>=0.8.0
httpx>=0.24.0
x402>=0.1.0